update the types

+++ contracts/cheddar_5x5_tic_tac_toe/src/utils.rs
@@ -21,7 +21,7 @@ pub(crate) type GameId = u64;
 pub(crate) type AffiliateId = AccountId;
 
 /// This constant can be used to set the board size
-pub(crate) const BOARD_SIZE: usize = 5;
+pub(crate) const BOARD_SIZE: u8 = 5;
 pub(crate) const MAX_NUM_TURNS: u64 = 25;
 pub(crate) const PLAYERS_NUM: usize = 2;
